+++ src/Namespace/Payments.php
@@ -76,10 +76,10 @@ interface Payments
     /**
      * Generate an [invoice deep link](https://core.telegram.org/api/links#invoice-links).
      *
-     * @param string|array $invoice_media Invoice @see https://docs.madelineproto.xyz/API_docs/types/InputMedia.html
+     * @param \danog\MadelineProto\EventHandler\Media|string|array $invoice_media Invoice @see https://docs.madelineproto.xyz/API_docs/types/InputMedia.html
      * @return array{_: 'payments.exportedInvoice', url: string} @see https://docs.madelineproto.xyz/API_docs/types/payments.ExportedInvoice.html
      */
-    public function exportInvoice(array|string $invoice_media): array;
+    public function exportInvoice(\danog\MadelineProto\EventHandler\Media|array|string $invoice_media): array;
 
     /**
      * Informs server about a purchase made through the App Store: for official applications only.