+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 MadelineProto was updated (8.0.0-beta114)!
 
 Features:
+- You can now get direct download links for or **directly download** stories, here's an example:
+
+- Add support for `parse_mode` parsing for story methods.
+- `getReply` now simply returns null if the message doesn't reply to any other message.
+- `getReply` now has an optional parameter that can be used to filter the returned message type.
+- Added `isUser()`, `isBot()` messages to check whether the current user is a user or a bot.
+- Improved IDE typehinting.
+- CLI bots: you can now optionally specify a default download link URL (used by `getDownloadLink`) in the settings.
+
+MadelineProto was updated (8.0.0-beta114)!
+
+Features:
 - Added `DialogMessagePinned` service message with a `getPinnedMessage()` method.
 
 Fixes:
@@ -9,6 +21,7 @@ Fixes:
 - Fixed startAndLoopMulti
 - Tweaked the default drop timeout on media DCs to avoid timeout errors on slow networks.
 - Now the admin list only contains user report peers.
+- Make `markdownEscape` method accessible
 
 ---